Decomposing the monolith into modules
Broke out the api callers for LaunchDarkly and StatSig into their own
modules.

Moved the logic for extracting, transforming, and loading data into
their own modules to make the main index file more readable.

Converted files to use ES6 module syntax to prepare for easier use
of ES6 libraries.
