+++ .env.example
@@ -1,0 +1,5 @@
+# This api key should be able to be used across LaunchDarkly projects
+LAUNCHDARKLY_API_KEY=api-xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
+
+# This api key controls which StatSig project the flags will be migrated into
+STATSIG_API_KEY=console-xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
