+++ lib/constants.mjs
@@ -1,0 +1,3 @@
+// This tag will be placed on all of your Statsig gates that are imported
+// It will be used to clean before migration runs and can be used for your internal accounting
+export const TAG_NAME = 'Migration Script';
