+++ package.json
@@ -1,6 +1,7 @@
 {
   "dependencies": {
     "axios": "^1.6.8",
-    "csv-writer": "^1.6.0"
+    "csv-writer": "^1.6.0",
+    "dotenv": "^16.4.7"
   }
 }
